+++ packages/auto_updater/lib/src/auto_updater.dart
@@ -64,6 +64,9 @@ class AutoUpdater {
         case 'before-quit-for-update':
           listener.onUpdaterBeforeQuitForUpdate(appcastItem);
           break;
+        case 'update-cancelled':
+          listener.onUpdaterUpdateCancelled();
+          break;
       }
     }
   }